Sanely handle Postgres connection errors
As is, if something goes wrong with the Postgres connection (i.e. the
server is restarted), Ruote will get stuck and continue to spew out
PG::UnableToSend errors until it is restarted.

This changes that to 'fail fast', so if a connection error happens now
it issues an abort to crash the Ruby process (with the exception and
backtrace). I admit it's not the cleanest way, but Ruote doesn't have
any logic to handle the storage engine failing. The intention is you
have something else supervising the Ruote process, that will restart it
if it crashes.

The following errors are treated as a connection error, and will cause
the Ruote process to crash:

* PG::AdminShutdown
* PG::UnableToSend
* PG::ConnectionBad

A new configuration option `abort_on_connection_error` has been added,
it defaults to `true`. If you have your own error handling code, to
maintain the legacy behaviour set it to `false` when initialising the
storage engine.